influxdata: add Value and test against line-protocol corpus
The Value type enables us to return field values without incurring
allocation overhead. With this in place, we're able to run the
tests against all the tests in the line-protocol corpus repository.
Because there's no good definition of "correct" yet, we define
a result to be correct if it matches the result from at least one current implementation.