rename influxdata to lineprotocol
Given that there will be other formats for writing data to InfluxDB,
"influxdata" no longer seems like a good choice of name.
Rename the package to the more specific `lineprotocol` and
also add more documentation. After this, we should be ready
to tag v2.0.0.

Also improve the docs somewhat and add a couple of examples.