lineprotocol: improve error messages
Previously, it would report "empty tag key" when there was actually
a tag key present. This improves the error message in that case
and a couple of others.

It also includes the field name and the field value when there's a field value error

It also returns timestamp errors with correct source locations
and the offending timestamp value included.
That required the tests to be changed to invoke `Time` rather
than `TimeBytes`, but that seems like a good thing and enables
us to test more cases.